refactor: update Next.js config to only build pages with a .page suffix
This allows us to keep .test files co-located with the pages/components that they're testing. Without this adjustment, Next.js'  build process fails.
