+++ pages/index.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Home from '.';
+import Home from './index.page';
 
 describe('<Home />', () => {
   it('should render without crashing', async () => {
