+++ pages/profile/[id]/index.page.tsx
